+++ src/Attributes/Response200Attribute.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
-namespace PowerUtils.AspNetCore.Extensions.Attributes
+namespace PowerUtils.AspNetCore.Attributes
 {
     public class Response200Attribute : ProducesResponseTypeAttribute
     {